requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
tenacity>=8.2.3